main_folder_backup/requirements.txt
aiohttp==3.9.3
base58==2.1.1
cryptography==44.0.3
protobuf==6.30.2
python-dotenv==1.1.0
python-telegram-bot==22.0
qrcode==8.2
Requests==2.32.3
solana==0.36.6
solders==0.26.0
